test(pacquet): move git runner tests
